Split models.py in two files
